Hoàn thiện chức năng của giáo vụ

+++ ltudjava_bt2/src/hibernate/dao/StudentDAO.java
@@ -29,8 +29,7 @@ public class StudentDAO {
         } catch (Exception e) {
             transaction.rollback();
             return false;
-        }
-        finally {
+        } finally {
             session.close();
         }
         return true;
@@ -38,15 +37,16 @@ public class StudentDAO {
 
     public static boolean removeStudent(String id) {
         Session session = HibernateUtil.getSessionFactory().openSession();
-        Student student = null;
+        Student student = getStudent(id);
+        if (student == null) return false;
+        Transaction transaction = null;
         try {
-            student = session.get(Student.class, id);
+            transaction = session.beginTransaction();
             session.delete(student);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return false;
-        }
-        finally {
+            transaction.commit();
+        } catch (Exception ex) {
+            ex.printStackTrace();
+        } finally {
             session.close();
         }
         return true;
@@ -71,21 +71,20 @@ public class StudentDAO {
         return true;
     }
 
-    public static List<Student> getList(String query){
+    public static List<Student> getList(String query) {
         Session session = HibernateUtil.getSessionFactory().openSession();
         List<Student> list = null;
         try {
             list = session.createQuery(query).list();
         } catch (Exception e) {
             e.printStackTrace();
-        }
-        finally {
+        } finally {
             session.close();
         }
         return list;
     }
 
-    public static List<Student> getList(){
+    public static List<Student> getList() {
         return getList("from hibernate.java.Student");
     }
 }
